requirements.txt
# Core web frameworks
streamlit>=1.24.0
streamlit-option-menu>=0.4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data processing
pandas>=1.5.3
numpy>=1.24.2
scikit-learn>=1.2.2
joblib>=1.2.0

# API and networking
requests>=2.28.2
httpx>=0.25.0

# Visualization
plotly>=5.13.0
matplotlib>=3.7.0

# Configuration and environment
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Time and timezone
pytz>=2023.3

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Additional utilities
watchdog>=3.0.0
click>=8.1.0 